move hardcoded shared prefs to standard prefs for sync with existing system
